Revert "refactor(libs): move libs folder to libs package (#483)"
This reverts commit 251ec9279f75394822db382ce6d1987f3c8b0d3d.
