+++ packages/app/src/components/PortfolioOverview.tsx
@@ -3,8 +3,9 @@ import styled from 'styled-components/macro';
 import { transparentize } from 'polished';
 import dayjs from 'dayjs';
 import { Select, Type } from '@tuja/components';
-import { formatCurrency, updatePortfolioName } from '@tuja/libs';
+import { formatCurrency } from '@tuja/libs';
 import EditableTitle from 'components/EditableTitle';
+import { updatePortfolioName } from 'libs/portfolio';
 import usePortfolio from 'hooks/usePortfolio';
 import usePortfolioProcessor from 'hooks/usePortfolioProcessor';
 import useStartDate from 'hooks/useStartDate';
