+++ packages/app/src/workers/processor.worker.ts
@@ -4,16 +4,9 @@ import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
 import minMax from 'dayjs/plugin/minMax';
 import firebase from 'firebase/app';
 import 'firebase/functions';
-import {
-  normalizeForex,
-  getForexPair,
-  TimeSeries,
-  StockHistory,
-  StockInfo,
-  StockLivePrice,
-  Portfolio,
-  PortfolioPerformance,
-} from '@tuja/libs';
+import { normalizeForex, getForexPair, TimeSeries } from '@tuja/libs';
+import type { Portfolio, PortfolioPerformance } from 'libs/portfolio';
+import { StockHistory, StockInfo, StockLivePrice } from 'libs/stocksClient';
 import {
   getActivitiesIterator,
   iterateActivities,
