+++ packages/libs/src/index.ts
@@ -1,5 +1,3 @@
 export * from './forex';
 export * from './timeSeries';
 export * from './activities';
-export * from './stocksData';
-export * from './portfolio';
